splitting views
